+++ isitfit/cost/optimizerListener.py
@@ -135,12 +135,15 @@ class OptimizerListener:
     return 'Underused', out_c2
 
 
-  def handle_pre(self):
+  def handle_pre(self, context_pre):
       # a csv file handle to which to stream results
       self.csv_fn_intermediate = tempfile.NamedTemporaryFile(prefix='isitfit-intermediate-', suffix='.csv', delete=False)
       logger.info(colored("Intermediate results will be streamed to %s"%self.csv_fn_intermediate.name, "cyan"))
       self.csv_fh = open(self.csv_fn_intermediate.name, 'w')
       self.csv_writer = csv.writer(self.csv_fh)
+
+      # done
+      return context_pre
 
 
   def per_ec2(self, context_ec2):